feat(ideaBox): refactor navigation and add IdeaBox and MySuggestions pages
